Services fail to start in strict mode (#1031)
- Per Mohit's suggestion, migrated the initService() piece inside run() so that we do the initialization on thread run after the certs are installed vs on object creation. Now only DefaultService calls initService(), not in subclasses like KafkaService/CassandraService.
- Removed try-catch around Throwable in initService(). If something throws an exception in there, we should catch the most specific one we can.
- Removed unneeded ElasticService class
- Removed some intermediate instance variables
- Refactored common Scheduler behavior into AbstractScheduler
