--- frameworks/elastic/src/main/java/com/mesosphere/sdk/elastic/scheduler/ElasticService.java
@@ -1,17 +1,0 @@
-package com.mesosphere.sdk.elastic.scheduler;
-
-import com.mesosphere.sdk.scheduler.SchedulerFlags;
-import com.mesosphere.sdk.specification.DefaultService;
-
-import java.io.File;
-
-/**
- * Service for the Elastic framework.
- */
-public class ElasticService extends DefaultService {
-
-    ElasticService(File pathToYamlSpecification) throws Exception {
-        super(pathToYamlSpecification, SchedulerFlags.fromEnv());
-    }
-
-}
