+++ sdk/scheduler/src/main/java/com/mesosphere/sdk/scheduler/AbstractScheduler.java
@@ -1,0 +1,130 @@
+package com.mesosphere.sdk.scheduler;
+
+import com.google.protobuf.TextFormat;
+import com.mesosphere.sdk.reconciliation.DefaultReconciler;
+import com.mesosphere.sdk.state.StateStore;
+import com.mesosphere.sdk.state.StateStoreUtils;
+import org.apache.mesos.Protos;
+import org.apache.mesos.Scheduler;
+import org.apache.mesos.SchedulerDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.concurrent.atomic.AtomicBoolean;
+
+/**
+ * Abstract {@link Scheduler} that provides some default behaviors around Mesos lifecycle events such as
+ * getting registered, reregistered, disconnected, etc.
+ */
+public abstract class AbstractScheduler implements Scheduler {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractScheduler.class);
+    protected final StateStore stateStore;
+    // Mesos may call registered() multiple times in the lifespan of a Scheduler process, specifically when there's
+    // master re-election. Avoid performing initialization multiple times, which would cause resourcesQueue to be stuck.
+    private final AtomicBoolean isAlreadyRegistered = new AtomicBoolean(false);
+    protected SchedulerDriver driver;
+    protected DefaultReconciler reconciler;
+
+    /**
+     * Creates a new AbstractScheduler given a {@link StateStore}.
+     */
+    protected AbstractScheduler(StateStore stateStore) {
+        this.stateStore = stateStore;
+    }
+
+    @Override
+    public void registered(SchedulerDriver driver, Protos.FrameworkID frameworkId, Protos.MasterInfo masterInfo) {
+        if (isAlreadyRegistered.getAndSet(true)) {
+            // This may occur as the result of a master election.
+            LOGGER.info("Already registered, calling reregistered()");
+            reregistered(driver, masterInfo);
+            return;
+        }
+
+        LOGGER.info("Registered framework with frameworkId: {}", frameworkId.getValue());
+        try {
+            this.reconciler = new DefaultReconciler(stateStore);
+            initialize(driver);
+        } catch (InterruptedException e) {
+            LOGGER.error("Initialization failed with exception: ", e);
+            SchedulerUtils.hardExit(SchedulerErrorCode.INITIALIZATION_FAILURE);
+        }
+
+        try {
+            stateStore.storeFrameworkId(frameworkId);
+        } catch (Exception e) {
+            LOGGER.error(String.format(
+                    "Unable to store registered framework ID '%s'", frameworkId.getValue()), e);
+            SchedulerUtils.hardExit(SchedulerErrorCode.REGISTRATION_FAILURE);
+        }
+
+        this.driver = driver;
+        postRegister();
+    }
+
+    protected abstract void initialize(SchedulerDriver driver) throws InterruptedException;
+
+    @Override
+    public void reregistered(SchedulerDriver driver, Protos.MasterInfo masterInfo) {
+        LOGGER.info("Re-registered with master: {}", TextFormat.shortDebugString(masterInfo));
+        postRegister();
+    }
+
+    @Override
+    public void offerRescinded(SchedulerDriver driver, Protos.OfferID offerId) {
+        LOGGER.warn("Ignoring rescinded Offer: {}.", offerId.getValue());
+    }
+
+    @Override
+    public void frameworkMessage(SchedulerDriver driver,
+                                 Protos.ExecutorID executorId,
+                                 Protos.SlaveID slaveId,
+                                 byte[] data) {
+        LOGGER.error("Received a Framework Message, but don't know how to process it");
+    }
+
+    @Override
+    public void disconnected(SchedulerDriver driver) {
+        LOGGER.error("Disconnected from Master.");
+        SchedulerUtils.hardExit(SchedulerErrorCode.DISCONNECTED);
+    }
+
+    @Override
+    public void slaveLost(SchedulerDriver driver, Protos.SlaveID agentId) {
+        // TODO: Add recovery optimizations relevant to loss of an Agent.  TaskStatus updates are sufficient now.
+        LOGGER.warn("Agent lost: {}", agentId.getValue());
+    }
+
+    @Override
+    public void executorLost(SchedulerDriver driver, Protos.ExecutorID executorId, Protos.SlaveID slaveId, int status) {
+        // TODO: Add recovery optimizations relevant to loss of an Executor.  TaskStatus updates are sufficient now.
+        LOGGER.warn("Lost Executor: {} on Agent: {}", executorId.getValue(), slaveId.getValue());
+    }
+
+    @Override
+    public void error(SchedulerDriver driver, String message) {
+        LOGGER.error("SchedulerDriver failed with message: " + message);
+
+        SchedulerUtils.hardExit(SchedulerErrorCode.ERROR);
+    }
+
+    void suppress() {
+        LOGGER.info("Suppressing offers.");
+        driver.suppressOffers();
+        StateStoreUtils.setSuppressed(stateStore, true);
+    }
+
+    void revive() {
+        LOGGER.info("Reviving offers.");
+        driver.reviveOffers();
+        StateStoreUtils.setSuppressed(stateStore, false);
+    }
+
+    protected void postRegister() {
+        reconciler.start();
+        reconciler.reconcile(driver);
+        revive();
+    }
+
+}
